pyproject.toml
[tool.poetry]
name = "ticketshop"
version = "0.1.0"
description = "Web API for online ticket shop"
authors = ["Zayats Roman"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
cryptography = "^46.0.2"
faststream = {extras = ["rabbit"], version = "^0.6.2"}
sqlmodel = "^0.0.25"
pyjwt = "^2.10.1"
pydantic = {extras = ["email"], version = "^2.11.10"}
asyncmy = "^0.2.10"
python-multipart = "^0.0.20"
anyio = "^4.11.0"
pytest-asyncio = "^1.2.0"
aiohttp = "^3.13.0"
docker = "^7.1.0"
pydantic-settings = "^2.13.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"